Refactor and enhance game dialog, inventory, and item management
- Updated struct.h to include a new function for struct selection.
- Refactored dialog.cpp to improve readability and maintainability.
- Changed inventory management to use selected_item instead of item.
- Updated inventory.h to reflect changes in variable names.
- Improved message handling in msg.cpp to allow for better input handling.
- Adjusted pause menu to use INT_MIN for option adder.
- Enhanced settings.cpp to include a restart confirmation dialog.
- Added item management functionality in item.cpp and itemup.cpp.
- Introduced new item handling functions for healing and food consumption.
- Created new data files for world configurations.
- Added error handling for missing item data.
- Improved overall code structure and consistency across files.

+++ windows/dynamic/pause.cpp
@@ -54,7 +54,7 @@ void pause(){
         "Settings",
         "Exit",
     };
-    option_adder(wino.top().screen["pau"] , opts , wino.top().stl["selecter"] , LLONG_MIN , LLONG_MIN , 1);
+    option_adder(wino.top().screen["pau"] , opts , wino.top().stl["selecter"] , INT_MIN , INT_MIN , 1);
     ita(wino.top().screen["screen"] ,wino.top().screen["pau"] , INT_MIN , INT_MIN , 1);
     itt();
 }
